+++ p1/AEngine/src/androidTest/java/es/ucm/vdm/aengine/AndroidInput.java
@@ -1,0 +1,13 @@
+package es.ucm.vdm.aengine;
+
+import android.view.MotionEvent;
+import android.view.View;
+
+import es.ucm.vdm.engine.Input;
+
+public class AndroidInput extends Input implements View.OnTouchListener {
+    @Override
+    public boolean onTouch(View view, MotionEvent motionEvent) {
+        return false;
+    }
+}
